Split more examples
